# order/impl/src/main/java/dev/thiagooliveira/delivery/orders/model/OrderEvent.java
package dev.thiagooliveira.delivery.orders.model;

import jakarta.persistence.*;
import java.time.LocalDateTime;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class OrderEvent {

    @EmbeddedId
    private OrderEventId id;

    @ManyToOne
    @MapsId("orderId")
    @JoinColumn(name = "order_id", nullable = false)
    private Order order;

    private LocalDateTime dateTime;
}